+++ src/statphys/input/Input_time.cpp
@@ -21,15 +21,15 @@ namespace simol
 
   /// Number of steps in the "output" part of the simulation
   /// Truncated to a multiple of the timeStep
-  int Input::nbOfSteps() const
+  long int Input::nbOfSteps() const
   {
     static bool warningGiven = false;
     if (data["Time"]["TotalNbOfSteps"])
-      return data["Time"]["TotalNbOfSteps"].as<int>();
+      return data["Time"]["TotalNbOfSteps"].as<long int>();
     else if (data["Time"]["TotalTime"])
     {
       double totalTime = data["Time"]["TotalTime"].as<double>();
-      int totalNb = (int)( totalTime / timeStep());
+      long int totalNb = totalTime / timeStep();
       double totalTrunc = totalNb * timeStep();
       if (!warningGiven && (totalTime != totalTrunc))
       {
@@ -43,15 +43,15 @@ namespace simol
 
   /// Number of steps in the thermalization part (temperatures imposed everywhere)
   /// Truncated to a multiple of the timeStep
-  int Input::thermalizationNbOfSteps() const
+  long int Input::thermalizationNbOfSteps() const
   {
     static bool warningGiven = false;
     if (data["Time"]["ThermalizationNbOfSteps"])
-      return data["Time"]["ThermalizationNbOfSteps"].as<int>();
+      return data["Time"]["ThermalizationNbOfSteps"].as<long int>();
     else if (data["Time"]["ThermalizationTime"])
     {
       double thermalTime = data["Time"]["ThermalizationTime"].as<double>();
-      int thermalNb = (int)( thermalTime / timeStep());
+      long int thermalNb = thermalTime / timeStep();
       double thermalTrunc = thermalNb * timeStep();
       if (!warningGiven && (thermalTime != thermalTrunc))
       {
@@ -66,15 +66,15 @@ namespace simol
 
   /// Number of steps in the burnIn part (no output)
   /// Truncated to a multiple of the timeStep
-  int Input::burninNbOfSteps() const
+  long int Input::burninNbOfSteps() const
   {
     static bool warningGiven = false;
     if (data["Time"]["BurnInNbOfSteps"])
-      return data["Time"]["BurnInNbOfSteps"].as<int>();
+      return data["Time"]["BurnInNbOfSteps"].as<long int>();
     else if (data["Time"]["BurnInTime"])
     {
       double burninTime = data["Time"]["BurnInTime"].as<double>();
-      int burninNb = (int)( burninTime / timeStep());
+      long int burninNb = burninTime / timeStep();
       double burninTrunc = burninNb * timeStep();
       if (!warningGiven && (burninTime != burninTrunc))
       {
